register/login layouts basic layouts done

+++ mud/src/components/Register.js
@@ -2,43 +2,45 @@ import React from 'react';
 import axios from 'axios';
 import { Link } from 'react-router-dom';
 
-class Login extends React.Component {
+class Register extends React.Component {
 	constructor() {
 		super();
 		this.state = {
 			username: '',
-			password: ''
+			password: '',
+			password2: ''
 		};
 	}
-
 	handleInput = (e) => {
 		this.setState({
 			[e.target.name]: e.target.value
 		});
 	};
 
-	handleSubmit = (e) => {
+	register = (e) => {
 		e.preventDefault();
 		const URL = '';
 		axios
 			.post(URL, {
 				username: this.state.username,
-				password: this.state.password
+				password: this.state.password,
+				confirmPassword: this.state.confirmPassword
 			})
 			.then((res) => {
 				localStorage.setItem('token');
-				// clears form
 				this.setState({
 					username: '',
-					password: ''
+					password: '',
+					password2: ''
 				});
-				this.PaymentResponse.history.push('');
+				this.props.history.push('/');
 			})
 			.catch((err) => {
-				alert('Enter Username and Password');
+				alert('Passwords must match');
 				this.setState({
 					username: '',
-					password: ''
+					password: '',
+					password2: ''
 				});
 				console.log(err);
 			});
@@ -46,9 +48,9 @@ class Login extends React.Component {
 
 	render() {
 		return (
-			<div className="login">
-				<h1 className="title">Run Fun</h1>
-				<form className="form" onSubmit={this.handleSubmit}>
+			<div className="register">
+				<h1>Register Here</h1>
+				<form className="form">
 					<label>Username</label>
 					<input
 						name="username"
@@ -65,15 +67,23 @@ class Login extends React.Component {
 						onChange={this.handleInput}
 						value={this.state.password}
 					/>
-					<p>
-						Don't have an account?
-						<Link to="/register">Register</Link>
-					</p>
-					<button>Sign In</button>
+					<label>Password</label>
+					<input
+						name="password2"
+						type="text"
+						placeholder="Confirm password"
+						onChange={this.handleInput}
+						value={this.state.password2}
+					/>
+					<button>Register</button>
 				</form>
+				<p>
+					Have an account?
+					<Link to="/">Sign In</Link>
+				</p>
 			</div>
 		);
 	}
 }
 
-export default Login;
+export default Register;
